+++ NEWS.md
@@ -6,6 +6,11 @@
 * Changed `SLIM_HOME` environmental variable to `SLIMR_PATH`, which now should be
 set with the full path to the SLiM executable, not just the directory containing it
 * Updated `slim_setup()` function which now provides cross-platform automatic installation using conda via the `reticulate` package. Now it should not be necessary for users to manually install SLiM any more
+* Simplified `slim_run()` API and temporarily removed ability to use R callbacks (the plan is to eventually replace this functionality with a new '`slimr` verb' function)
+* Added new '`slimr` verb' `r_message()`. `r_message()` can be called in a `slim_block()` to print a message to the R console while the SLiM simulation is running. This is useful for debugging and monitoring progress of a simulation.
+* `slim_run()` with `progress = TRUE` now produces a progress bar for any simulation (previously it only worked if you created output with `r_output()`) and is much more efficient (produces less overhead)
+* `slim_script_render_dots()` function added. It works just like `slim_script_render()`, but accepts values to be filled in a templated `slimr_script` as named arguments. For example `slim_script_render_dots(x, mut_rate = c(1e-6, 1e-7))` will render the `slimr_script` called `x` with `mut_rate` set to `1e-6` and `1e-7`. This can sometimes be less verbose than using a list or data frame to fill in a templated `slimr_script`, especially useful when you simply want to test different values quickly.
+
 
 # slimr 0.3.0